Fixing bug when a river reached the first cell of domain
The software sets that a side is close if there are no water cells
on that side. If a river starts from the first cell of the domain,
though, the software may mistake the first cell for the river for
an open boundary. Now this is fixed by checkign for the openess
of the boundaries before digging the rivers.